move wizard button to menu
