+++ src/main/java/controllers/MainUIController.java
@@ -4,7 +4,11 @@ import javafx.application.Platform;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import model.data.DataModel;
+import view.Dialog;
+import view.XMLCreationDialog;
 
+import java.io.IOException;
+import java.util.logging.Level;
 import java.util.logging.Logger;
 
 /**
@@ -13,6 +17,9 @@ import java.util.logging.Logger;
  *
  */
 public class MainUIController {
+
+	private static final double WIZARD_DIALOG_WIDTH = 1000;
+	private static final double WIZARD_DIALOG_HEIGHT = 640;
 
 	@FXML private TableViewController tableViewController;
 	@FXML private DataController dataController;
@@ -45,7 +52,17 @@ public class MainUIController {
 		analysisController.setDataModel(model);
 		visualizationController.setModel(model);
 		visualizationController.initializeVisualisation();
+	}
 
+	@FXML
+	protected void startWizard(ActionEvent actionEvent) {
+		try {
+			Dialog wizardDialog = new XMLCreationDialog();
+			wizardDialog.setSize(WIZARD_DIALOG_WIDTH, WIZARD_DIALOG_HEIGHT);
+			wizardDialog.show();
+		} catch (IOException e) {
+			logger.log(Level.SEVERE, "FXML error: " + e.getMessage());
+		}
 	}
 
 }
