+++ src/main/resources/tab_data.fxml
@@ -19,8 +19,6 @@
 						onAction="#handleImportButtonAction"
 						text="Import" />
 				<Label layoutX="14.0" layoutY="7.0" text="Import data" />
-
-				<Button onAction="#startWizard" text="XML Creation Wizard"/>
 			</children>
 		</AnchorPane>
 	</content>
